// types.ts
export enum Screen {
  Home = 'Inicio',
  Training = 'Entreno',
  Rest = 'Descanso',
  Classes = 'Clases',
  Profile = 'Perfil'
}

export enum Role {
  User = 'User',
  Admin = 'Admin'
}

export enum AdminScreenType {
  Dashboard = 'Dashboard',
  Members = 'Gestión de Miembros',
  Schedule = 'Gestión de Horarios y Clases',
  Reports = 'Reportes y Analíticas',
  Notifications = 'Sistema de Notificaciones'
}

export enum ProfileSubScreen {
  Main = 'Main',
  Edit = 'Edit',
  Notifications = 'Notifications'
}